Changing epublib JAR file.
I'm returning the old epublib-core-latest.jar file that was in the original epub3reader project. Inside the JAR file itself it says this is version 3.1, which is the same version as the JAR file I replaced it by. The one I compiled from the provided sources from the author's github page.

I'm returning the old one (but now under the name epublib-core-3.1.jar) because it is:
a) smaller (216KB vs. 277KB)
b) works faster (especially on older devices like my Nook Tablet)

Reason for that is probably minimization of the Java code in the JAR file.