Switched from supporting Support Library v4 to v7-appcompat.
It gives us support of ActionBarDrawerToggle necessary to create a proper navigation drawer (other method is deprecated).

However, because it contains resources, is necessary to add it as a library project into Eclipse, and then add it as a library in the eclipse project settings of our application.

Added following this manual:
  Android -> Support Library Setup -> Adding libraries with resources -> Eclipse
  http://developer.android.com/tools/support-library/setup.html#libs-with-res